Restructured entire project
