+++ .flake8
@@ -1,0 +1,4 @@
+[flake8]
+ignore = E117,E501,E722,W191
+exclude =
+    tests
