+++ .github/workflows/flake8.yml
@@ -8,18 +8,14 @@ jobs:
     steps:
       - name: Checkout repo
         uses: actions/checkout@v2
-      - name: Set up Python 3.7
+      - name: Set up Python 3.8
         uses: actions/setup-python@v2
         with:
-          python-version: 3.7
+          python-version: 3.8
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
           pip install -r requirements.txt
       - name: Lint with flake8
         run: |
-          pip install flake8
-          flake8 --ignore=E117,E501,E722,W191 jeeves.py
-          flake8 --ignore=E117,E501,E722,W191 report.py
-          flake8 --ignore=E117,E501,E722,W191 remind.py
-          flake8 --ignore=E117,E501,E722,W191 functions.py
+          flake8
