+++ requirements.txt
@@ -5,3 +5,4 @@ python-bugzilla==3.0.2
 jira==2.0.0
 pytest==6.2.2
 pytest-cov==2.11.1
+flake8==3.8.4
