+++ tests/test_common.py
@@ -1,0 +1,19 @@
+from jeeves.common import *
+
+
+def test_generate_header():
+	pass
+
+
+def test_generate_html_file():
+	pass
+
+
+def test_percent_func():
+	assert percent(0, 1) == 0.0
+	assert percent(1, 2) == 50.0
+	assert percent(1, 1) == 100.0
+
+
+def test_validate_config():
+	pass
