+++ tests/test_jobs.py
@@ -1,25 +1,4 @@
-from functions import *
-
-
-def test_generate_header():
-	pass
-
-
-def test_generate_html_file():
-	pass
-
-
-def get_bugs_dict():
-	pass
-
-
-def test_get_bugs_set():
-	mockers = {
-		'job1': {'bz': [0]},
-		'job2': {'bz': [123456]},
-		'job3': {'bz': [123456, 789123]}
-	}
-	assert get_bugs_set(mockers) == {123456, 789123}
+from jeeves.jobs import *
 
 
 def test_get_jenkins_job_info():
@@ -28,19 +7,6 @@ def test_get_jenkins_job_info():
 
 def test_get_jenkins_jobs():
 	pass
-
-
-def test_get_tickets_dict():
-	pass
-
-
-def test_get_tickets_set():
-	mockers = {
-		'job1': {'jira': [0]},
-		'job2': {'jira': ['RHOSINFRA-123']},
-		'job3': {'jira': ['RHOSINFRA-123', 'RHOSENTDFG-456']}
-	}
-	assert get_tickets_set(mockers) == {'RHOSINFRA-123', 'RHOSENTDFG-456'}
 
 
 def test_get_osp_version():
@@ -59,40 +25,3 @@ def test_get_osp_version():
 	assert get_osp_version('DFG-security-keystone-16_director-rhel-virthost-1cont_1comp-ipv4-geneve-lvm-containers') == '16'
 	assert get_osp_version('DFG-upgrades-updates-16-from-passed_phase1-HA-ipv4') == '16'
 	assert get_osp_version('DFG-all-unified-weekly-multijob') is None
-
-
-def test_get_other_blockers():
-	pass
-
-
-def test_has_blockers():
-	mockers = {
-		'job1': {'bz': [123456]},
-		'job2': {'jira': ['RHOSINFRA-123']},
-		'job3': {'other': {'name': 'this is a test name'}},
-		'job4': {'bz': [0]},
-		'job5': {'jira': [0]},
-		'job6': {'other': [0]},
-		'job7': {'owners': 'foo@bar.com'},
-		'job8': {'owners': 'foo@bar.com', 'bz': [123456], 'jira': ['RHOSINFRA-123']},
-		'job9': {}
-	}
-	assert has_blockers(mockers, 'job1') == True
-	assert has_blockers(mockers, 'job2') == True
-	assert has_blockers(mockers, 'job3') == True
-	assert has_blockers(mockers, 'job4') == False
-	assert has_blockers(mockers, 'job5') == False
-	assert has_blockers(mockers, 'job6') == False
-	assert has_blockers(mockers, 'job7') == False
-	assert has_blockers(mockers, 'job8') == True
-	assert has_blockers(mockers, 'job9') == False
-
-
-def test_percent_func():
-	assert percent(0, 1) == 0.0
-	assert percent(1, 2) == 50.0
-	assert percent(1, 1) == 100.0
-
-
-def test_validate_config():
-	pass
