+++ tests/test_remind.py
@@ -1,0 +1,5 @@
+from jeeves.remind import *
+
+
+def test_run_remind():
+    pass
