+++ tests/test_report.py
@@ -1,0 +1,5 @@
+from jeeves.report import *
+
+
+def test_run_report():
+    pass
